Move all classes to packages
