+++ .gitignore
@@ -15,4 +15,5 @@ bin/
 
 text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
+text-ui-test/shibaData/*
 shibaData/*
