+++ src/main/java/shiba/datetimeformats/DateOptionalTime.java
@@ -1,6 +1,6 @@
-package datetimeformats;
+package shiba.datetimeformats;
 
-import exceptions.InvalidCommandException;
+import shiba.exceptions.InvalidCommandException;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
