+++ src/main/java/shiba/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package shiba.exceptions;
 
 public class InvalidCommandException extends ShibaException {
     public InvalidCommandException(String msg) {
