+++ src/main/java/shiba/exceptions/ShibaException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package shiba.exceptions;
 
 public class ShibaException extends Exception {
     public ShibaException(String message) {
