+++ src/main/java/shiba/parsers/SpaceSeparatedValuesParser.java
@@ -1,4 +1,4 @@
-package parsers;
+package shiba.parsers;
 
 import java.util.ArrayList;
 import java.util.List;
